requirements.txt
Scrapy==2.11.0
pymongo==4.6.1
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
schedule==1.2.0
APScheduler==3.10.4
